refactor(frontend): change folder name bb-134
