+++ apps/frontend/src/pages/onboarding/libs/constants/constants.ts
