+++ apps/frontend/src/pages/onboarding/libs/constants/onboarding-form-default-values.contant.ts
